Remove unnecessary tests
