--- packages/hooks_riverpod/test/change_notifier_test.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  testWidgets('listen to the notifier', (tester) async {
-    final notifier = TestNotifier();
-    final provider = ChangeNotifierProvider((_) => notifier);
-
-    await tester.pumpWidget(
-      ProviderScope(
-        child: HookConsumer(builder: (c, ref, child) {
-          return Text(
-            ref.watch(provider).count.toString(),
-            textDirection: TextDirection.ltr,
-          );
-        }),
-      ),
-    );
-
-    expect(find.text('0'), findsOneWidget);
-
-    notifier.count++;
-    await tester.pump();
-
-    expect(find.text('1'), findsOneWidget);
-
-    await tester.pumpWidget(Container());
-
-    expect(notifier.mounted, isFalse);
-  });
-}
-
-class TestNotifier extends ChangeNotifier {
-  bool mounted = true;
-
-  int _count = 0;
-  int get count => _count;
-  set count(int count) {
-    _count = count;
-    notifyListeners();
-  }
-
-  @override
-  void dispose() {
-    mounted = false;
-    super.dispose();
-  }
-}
